[Frontend] Extract UserMenu component from library TopNav
